GPU IVFSQ code_size fix
Summary:
This bug was introduced in D24064745, which broke the code distance for GPU IVFSQ. The `code_size` is size in bytes per encoded vector, not per scalar. This diff updates the expression for computing GPU and CPU vector sizes.

The bug was not seen on FB machines since it appears that the memory allocator (jemalloc?) was more forgiving in terms of mapped page sizes, and garbage tends to be far away in N-dimensional space from real queries.

Reviewed By: beauby

Differential Revision: D24405231

fbshipit-source-id: f9ad0d3f326afe412ea864537a24efbd74d97f1f