Prevent snprintf vulnerability
Summary:
With a very big name for a `ParameterRange`, the `snprintf` call from `combination_name` can end up having a negative second parameter, causing  a memory overflow, which can lead to a serious security issue.

We can checking that the second parameter is always >= 0 and throw an exception if not.

See the new GTEST.

Reviewed By: mdouze

Differential Revision: D46856956

fbshipit-source-id: 91c657ec028c462d4b808b595811342034e00133